Switched to backing data as linked list, could not talk myself out of it

+++ src/main/java/com/jjoe64/graphview/Viewport.java
@@ -1369,7 +1369,7 @@ public class Viewport {
        mMinimalViewport.set(minX, maxY, maxX, minY);
     }
 
-    public void zoomToFit(float paddingTop, float paddingBottom, float paddingLeft, float paddingRight,
+    public void zoomToFit(double paddingTop, double paddingBottom, double paddingLeft, double paddingRight,
                           boolean xFromZero, boolean yFromZero) {
         if (xFromZero) {
             setMinX(0);
